HOTT-3372: Encapsulate all result collection interactions
