--- app/models/bulk_search.rb
@@ -1,55 +1,0 @@
-class BulkSearch
-  INITIAL_STATE = :queued
-  COMPLETE_STATE = :completed
-  FAILED_STATE = :failed
-  NOT_FOUND_STATE = :not_found
-  PROCESSING_STATE = :processing
-
-  STATE_MESSAGES = {
-    INITIAL_STATE => 'Your bulk search request has been accepted and is now on a queue waiting to be processed',
-    COMPLETE_STATE => 'Completed',
-    FAILED_STATE => 'Failed',
-    NOT_FOUND_STATE => 'Not found. Do you need to submit a bulk search request again? They expire in 2 hours',
-    PROCESSING_STATE => 'Processing',
-  }.freeze
-
-  HTTP_CODES = {
-    INITIAL_STATE => 202, # Accepted
-    PROCESSING_STATE => 202, # Accepted
-    COMPLETE_STATE => 200, # OK
-    FAILED_STATE => 500, # Internal Server Error
-    NOT_FOUND_STATE => 404, # Not Found
-  }.freeze
-  TWO_HOURS = 60 * 60 * 2
-
-  class << self
-    delegate :redis, to: TradeTariffBackend
-
-    def enqueue(searches)
-      ResultCollection.build(searches).tap do |result|
-        redis.set(
-          result.id,
-          Zlib::Deflate.deflate(result.to_json),
-          ex: TWO_HOURS,
-        )
-
-        BulkSearchWorker.perform_async(result.id)
-      end
-    end
-
-    def find(id)
-      json_blob = redis.get(id)
-
-      result = if json_blob.blank?
-                 {
-                   id:,
-                   status: NOT_FOUND_STATE,
-                 }
-               else
-                 JSON.parse(Zlib::Inflate.inflate(json_blob)).deep_symbolize_keys
-               end
-
-      ResultCollection.new(result)
-    end
-  end
-end
