+++ app/models/bulk_search/search_ancestor.rb
@@ -1,4 +1,4 @@
-class BulkSearch
+module BulkSearch
   class SearchAncestor
     include ContentAddressableId
