+++ app/workers/bulk_search_worker.rb
@@ -10,21 +10,8 @@ class BulkSearchWorker
     BulkSearchService.new(id).call
     Rails.logger.info("BulkSearchWorker: #{id} - #{BulkSearch.find(id)}")
   rescue StandardError => e
-    update_status(id, :error)
+    BulkSearch.find(id).failed!
 
     Rails.logger.error(e)
-  end
-
-  private
-
-  def update_status(id, status)
-    result = BulkSearch.find(id)
-    result.status = status
-
-    redis.set(
-      result.id,
-      Zlib::Deflate.deflate(result.to_json),
-      ex: BulkSearch::TWO_HOURS,
-    )
   end
 end
